+++ tiger-proxy/src/test/java/de/gematik/test/tiger/proxy/TestTigerProxyRouting.java
@@ -77,19 +77,20 @@ class TestTigerProxyRouting extends AbstractFastTigerProxyTest {
 
   public static Stream<Arguments> nestedAndShallowPathTestCases() {
     return Stream.of(
+        // fromPath, requestPath, actualPath, expectedReturnCode
         Arguments.of("/deep", "/foobar", "/deep/foobar", 777),
         Arguments.of("/deep", "/foobar/", "/deep/foobar/", 777),
         Arguments.of("/deep/", "/foobar", "/deep/foobar", 777),
         Arguments.of("/deep/", "/foobar/", "/deep/foobar/", 777),
-        // "/foobar,'','/foobar'", //TODO TGR-949: mockserver-bug. should work
-        Arguments.of("/foobar", "/", "/foobar/", 666),
+        Arguments.of("/foobar", "", "/foobar", 666), // 5
+        Arguments.of("/foobar", "/", "/foobar", 666),
         Arguments.of("/foobar/", "", "/foobar/", 666),
         Arguments.of("/foobar/", "/", "/foobar/", 666),
-        Arguments.of("", "/foobar", "/foobar", 666),
+        Arguments.of("", "/foobar", "/foobar", 666), // 9
         Arguments.of("", "/foobar/", "/foobar/", 666),
         Arguments.of("/", "/foobar", "/foobar", 666),
         Arguments.of("/", "/foobar/", "/foobar/", 666),
-        // "'','',''", //TODO TGR-949: mockserver-bug. should work
+        Arguments.of("", "", "/", 888), // 13
         Arguments.of("", "/", "/", 888),
         Arguments.of("/", "", "/", 888),
         Arguments.of("/", "/", "/", 888));